Extract navigation PID controller to separate module
